feat: Add table cell trimming for cqueue feat: Add table cell trimming for cacct feat: Add table cell trimming for cacctmgr
