+++ internal/util/util.go
@@ -17,12 +17,8 @@
 package util
 
 import (
-	"strings"
-
 	nested "github.com/antonfisher/nested-logrus-formatter"
-	"github.com/olekukonko/tablewriter"
 	log "github.com/sirupsen/logrus"
-	"google.golang.org/protobuf/types/known/durationpb"
 )
 
 type Config struct {
@@ -49,64 +45,6 @@ func init() {
 	DefaultConfigPath = "/etc/crane/config.yaml"
 	DefaultCforedServerListenAddress = "0.0.0.0"
 	DefaultCforedServerListenPort = "10012"
-}
-
-func SetBorderlessTable(table *tablewriter.Table) {
-	table.SetBorder(false)
-	table.SetAutoFormatHeaders(true)
-	table.SetAutoWrapText(false)
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT)
-	table.SetHeaderLine(false)
-	table.SetCenterSeparator("")
-	table.SetColumnSeparator("")
-	table.SetRowSeparator("")
-	table.SetTablePadding(" ")
-	table.SetNoWhiteSpace(true)
-}
-
-func SetBorderTable(table *tablewriter.Table) {
-	table.SetBorders(tablewriter.Border{Left: true, Top: true, Right: true, Bottom: true})
-	table.SetAutoFormatHeaders(true)
-	table.SetAutoWrapText(true)
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT)
-	table.SetCenterSeparator("|")
-	table.SetTablePadding("\t")
-}
-
-func FormatTable(tableOutputWidth []int, tableHeader []string,
-	tableData [][]string) (formatTableHeader []string, formatTableData [][]string) {
-	for i, h := range tableHeader {
-		if tableOutputWidth[i] != -1 {
-			padLength := tableOutputWidth[i] - len(h)
-			if padLength >= 0 {
-				tableHeader[i] = h + strings.Repeat(" ", padLength)
-			} else {
-				tableHeader[i] = h[:tableOutputWidth[i]]
-			}
-		}
-	}
-	for i, row := range tableData {
-		for j, cell := range row {
-			if tableOutputWidth[j] != -1 {
-				padLength := tableOutputWidth[j] - len(cell)
-				if padLength >= 0 {
-					tableData[i][j] = cell + strings.Repeat(" ", padLength)
-				} else {
-					tableData[i][j] = cell[:tableOutputWidth[j]]
-				}
-			}
-		}
-	}
-	return tableHeader, tableData
-}
-
-func InvalidDuration() *durationpb.Duration {
-	return &durationpb.Duration{
-		Seconds: 315576000000,
-		Nanos:   0,
-	}
 }
 
 func InitLogger(level log.Level) {
